% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propeller.R
\name{.extractSCE}
\alias{.extractSCE}
\title{Extract metadata from \code{SingleCellExperiment} object}
\usage{
.extractSCE(x)
}
\arguments{
\item{x}{object of class \code{SingleCellExperiment}}
}
\value{
a dataframe containing clusters, sample and group
}
\description{
This is an accessor function that extracts cluster, sample and group
information for each cell.
}
\author{
Belinda Phipson
}
