% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propeller.R
\name{.extractSeurat}
\alias{.extractSeurat}
\title{Extract metadata from \code{Seurat} object}
\usage{
.extractSeurat(x)
}
\arguments{
\item{x}{object of class \code{Seurat}}
}
\value{
a dataframe containing clusters, sample and group
}
\description{
This is an accessor function that extracts cluster, sample and group
information for each cell.
}
\author{
Belinda Phipson
}
