% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_pbmc_props.R
\name{load_pbmc_props}
\alias{load_pbmc_props}
\title{Get PBMC proportions data}
\usage{
load_pbmc_props()
}
\value{
a list object with cell type proportions, sample information and
total number of cells
}
\description{
Get PBMC proportions data
}
\examples{

pbmc_list <- load_pbmc_props()
names(pbmc_list)
pbmc_props <- pbmc_list$proportions
pbmc_sample_info <- pbmc_list$sample_info
tot_cells <- pbmc_list$total_cells

}
