% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normSca.R
\name{normSca}
\alias{normSca}
\title{Normalise and scale counts matrix}
\usage{
normSca(x, lib.size = lib.size, log = TRUE, prior.count = 0.5)
}
\arguments{
\item{x}{a matrix of counts}

\item{lib.size}{the library size}

\item{log}{logical, indicates whether the output should be on the log2 scale
or counts scale. Default is TRUE.}

\item{prior.count}{The prior count to add if the data is log2 normalised.
Default is a small count of 0.5.}
}
\value{
a dataframe of log-normalised and scaled counts
}
\description{
This function is called by the \code{preprocess} function and performs
log-normalisation and scaling of a counts matrix.
}
\examples{

y <- matrix(rnbinom(1000, size=2, mu= 20),ncol=10)
colnames(y)<- paste("Cell",1:10, sep="")
row.names(y)<-paste("Gene",1:100,sep="")
norm.data <- normSca(y,lib.size=colSums(y))

#Visualise the counts vs scaled data
boxplot(y)
boxplot(norm.data)

}
