% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speckle-package.R
\docType{package}
\name{speckle-package}
\alias{speckle}
\alias{speckle-package}
\title{speckle: Statistical methods for analysing single cell RNA-seq data}
\description{
The speckle package contains functions for the analysis of single cell RNA-seq data. The speckle package currently contains functions to analyse differences in cell type proportions. There are also functions to estimate the parameters of the Beta distribution based on a given counts matrix, and a function to normalise a counts matrix to the median library size. There are plotting functions to visualise cell type proportions and the mean-variance relationship in cell type proportions and counts. As our research into specialised analyses of single cell data continues we anticipate that the package will be updated with new functions.
}
\author{
\strong{Maintainer}: Belinda Phipson \email{phipson.b@wehi.edu.au}

}
\keyword{internal}
