% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speckle_example_data.R
\name{speckle_example_data}
\alias{speckle_example_data}
\title{Generate example data}
\usage{
speckle_example_data()
}
\value{
a dataframe containing cell-level information for sample, group and
clusters
}
\description{
Generate example data
}
\examples{

speckle_example_data()

}
