% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speckle-package.R
\docType{package}
\name{speckle-package}
\alias{speckle}
\alias{speckle-package}
\title{speckle: Statistical methods for analysing single cell RNA-seq data}
\description{
speckle contains functions for the analysis of single cell RNA-seq data.
}
\keyword{internal}
